#368: polished a bit
